Finall...

+++ LetsTogether/src/org/dimigo/service/ApplyService.java
@@ -23,6 +23,21 @@ public class ApplyService extends AbstractService {
 		}
 	}
 	
+	public void searchApplyAndUpdate(ApplyVO U) throws Exception {
+		Connection con = null;
+		try { 
+			con = getConnection();
+			
+			ApplyDao dao = new ApplyDao(con);
+			dao.searchApplyAndUpdate(U);
+			
+		} finally {
+			if(con != null) {
+				con.close();
+			}
+		}
+	}
+	
 	public List<ApplyVO> searchApplyList() throws Exception {
 		Connection con = null;
 		try { 
@@ -53,6 +68,36 @@ public class ApplyService extends AbstractService {
 		}
 	}
 	
+	public List<ApplyVO> searchApplyListByDoapplyid(String j) throws Exception {
+		Connection con = null;
+		try { 
+			con = getConnection();
+			
+			ApplyDao dao = new ApplyDao(con);
+			return dao.searchApplyListByDoapplyid(j);
+			
+		} finally {
+			if(con != null) {
+				con.close();
+			}
+		}
+	}
+	
+	public List<ApplyVO> searchApplyListByGetapplyid(String j) throws Exception {
+		Connection con = null;
+		try { 
+			con = getConnection();
+			
+			ApplyDao dao = new ApplyDao(con);
+			return dao.searchApplyListByGetapplyid(j);
+			
+		} finally {
+			if(con != null) {
+				con.close();
+			}
+		}
+	}
+	
 	public void doapply(ApplyVO user) throws Exception {
 		Connection con = null;
 		try { 
@@ -69,13 +114,15 @@ public class ApplyService extends AbstractService {
 		}
 	}
 	
-	public void signout(ApplyVO user) throws Exception {
+	public void deleteApply(ApplyVO user) throws Exception {
 		Connection con = null;
 		try { 
 			con = getConnection();
 			
 			ApplyDao dao = new ApplyDao(con);
-			
+			if(dao.searchApply(user) == null) {
+				throw new Exception("아직 신청하지 않았습니다.");
+			}
 			dao.deleteApply(user);
 		} finally {
 			if(con != null) {
@@ -83,4 +130,5 @@ public class ApplyService extends AbstractService {
 			}
 		}
 	}
+
 }
